♻️ [refactor] RecordingCardWithWebBridge 생성 (비즈니스 로직 포함된 RecordingCard)
