+++ apps/web/src/entities/record/ui/RecordingCard/index.ts
@@ -1,7 +1,6 @@
 import { Root } from './Root'
 import { Icon } from './Icon'
 import { Info } from './Info'
-import { IconWrapper } from './IconWrapper'
 
 /**
  * RecordingCard
@@ -29,5 +28,4 @@ export const RecordingCard = {
   Root,
   Icon,
   Info,
-  IconWrapper,
 }
